+++ crates/action/src/sync_node_project.rs
@@ -1,11 +1,11 @@
 use crate::action::{Action, ActionStatus};
 use crate::context::ActionContext;
 use crate::errors::ActionError;
-use moon_config::{NodeVersionFormat, TypeScriptConfig};
+use moon_config::{DependencyScope, NodeVersionFormat, TypeScriptConfig};
 use moon_lang_node::{package::PackageJson, tsconfig::TsConfigJson};
 use moon_logger::{color, debug};
 use moon_project::Project;
-use moon_utils::{fs, is_ci, path, string_vec};
+use moon_utils::{fs, is_ci, path, semver, string_vec};
 use moon_workspace::Workspace;
 use std::collections::{BTreeMap, HashSet};
 use std::path::Path;
@@ -95,10 +95,12 @@ pub async fn sync_node_project(
 
     // Sync each dependency to `tsconfig.json` and `package.json`
     let mut package_prod_deps: BTreeMap<String, String> = BTreeMap::new();
+    let mut package_peer_deps: BTreeMap<String, String> = BTreeMap::new();
+    let mut package_dev_deps: BTreeMap<String, String> = BTreeMap::new();
     let mut tsconfig_project_refs: HashSet<String> = HashSet::new();
 
-    for dep_id in project.get_dependencies() {
-        let dep_project = workspace.projects.load(&dep_id)?;
+    for dep_cfg in &project.dependencies {
+        let dep_project = workspace.projects.load(&dep_cfg.id)?;
         let dep_relative_path = path::to_virtual_string(
             path::relative_from(&dep_project.root, &project.root).unwrap_or_default(),
         )?;
@@ -112,21 +114,38 @@ pub async fn sync_node_project(
             if let Some(dep_package_json) = PackageJson::read(&dep_project.root)? {
                 if let Some(dep_package_name) = &dep_package_json.name {
                     let version_prefix = format.get_prefix();
+                    let dep_package_version = dep_package_json.version.unwrap_or_default();
                     let dep_version = match format {
                         NodeVersionFormat::File | NodeVersionFormat::Link => {
                             format!("{}{}", version_prefix, dep_relative_path)
                         }
                         NodeVersionFormat::Version
                         | NodeVersionFormat::VersionCaret
-                        | NodeVersionFormat::VersionTilde => format!(
-                            "{}{}",
-                            version_prefix,
-                            dep_package_json.version.unwrap_or_default()
-                        ),
+                        | NodeVersionFormat::VersionTilde => {
+                            format!("{}{}", version_prefix, dep_package_version)
+                        }
                         _ => version_prefix,
                     };
 
-                    package_prod_deps.insert(dep_package_name.to_owned(), dep_version);
+                    match dep_cfg.scope {
+                        DependencyScope::Production => {
+                            package_prod_deps.insert(dep_package_name.to_owned(), dep_version);
+                        }
+                        DependencyScope::Development => {
+                            package_dev_deps.insert(dep_package_name.to_owned(), dep_version);
+                        }
+                        DependencyScope::Peer => {
+                            // Peers are unique, so lets handle this manually here for now.
+                            // Perhaps we can wrap this in a new setting in the future.
+                            package_peer_deps.insert(
+                                dep_package_name.to_owned(),
+                                format!(
+                                    "^{}.0.0",
+                                    semver::extract_major_version(&dep_package_version)
+                                ),
+                            );
+                        }
+                    }
 
                     debug!(
                         target: LOG_TARGET,
@@ -150,18 +169,33 @@ pub async fn sync_node_project(
             debug!(
                 target: LOG_TARGET,
                 "Syncing {} as a project reference to {}'s {}",
-                color::id(&dep_id),
-                color::id(project_id),
+                color::id(&dep_project.id),
+                color::id(&project.id),
                 color::file(tsconfig_branch_name)
             );
         }
     }
 
     // Sync to the project's `package.json`
-    if !package_prod_deps.is_empty() {
+    if !package_prod_deps.is_empty()
+        || !package_dev_deps.is_empty()
+        || !package_peer_deps.is_empty()
+    {
         PackageJson::sync(&project.root, |package_json| {
             for (name, version) in package_prod_deps {
                 if package_json.add_dependency(&name, &version, true) {
+                    mutated_files = true;
+                }
+            }
+
+            for (name, version) in package_dev_deps {
+                if package_json.add_dev_dependency(&name, &version, true) {
+                    mutated_files = true;
+                }
+            }
+
+            for (name, version) in package_peer_deps {
+                if package_json.add_peer_dependency(&name, &version, true) {
                     mutated_files = true;
                 }
             }