+++ crates/config/src/lib.rs
@@ -7,6 +7,7 @@ mod validators;
 mod workspace;
 
 pub use errors::format_figment_errors;
+pub use project::dep::*;
 pub use project::global::GlobalProjectConfig;
 pub use project::task::*;
 pub use project::*;