+++ crates/project/tests/project_test.rs
@@ -1,6 +1,6 @@
 use moon_config::{
-    GlobalProjectConfig, PlatformType, ProjectConfig, ProjectLanguage, ProjectMetadataConfig,
-    ProjectType, TargetID, TaskConfig, TaskMergeStrategy, TaskOptionsConfig,
+    GlobalProjectConfig, PlatformType, ProjectConfig, ProjectDependsOn, ProjectLanguage,
+    ProjectMetadataConfig, ProjectType, TargetID, TaskConfig, TaskMergeStrategy, TaskOptionsConfig,
 };
 use moon_project::{Project, ProjectError};
 use moon_task::{EnvVars, FileGroup, Target, Task};
@@ -114,7 +114,7 @@ fn basic_config() {
         Project {
             id: String::from("basic"),
             config: ProjectConfig {
-                depends_on: string_vec!["noConfig"],
+                depends_on: vec![ProjectDependsOn::String("noConfig".to_owned())],
                 file_groups: HashMap::from([(String::from("tests"), string_vec!["**/*_test.rs"])]),
                 language: ProjectLanguage::JavaScript,
                 ..ProjectConfig::default()
@@ -185,7 +185,7 @@ fn overrides_global_file_groups() {
         Project {
             id: String::from("basic"),
             config: ProjectConfig {
-                depends_on: string_vec!["noConfig"],
+                depends_on: vec![ProjectDependsOn::String("noConfig".to_owned())],
                 file_groups: HashMap::from([(String::from("tests"), string_vec!["**/*_test.rs"])]),
                 language: ProjectLanguage::JavaScript,
                 ..ProjectConfig::default()